#84 Rename module (azel → api)
