+++ azely/__init__.py
@@ -2,7 +2,7 @@ __all__ = [
     "Location",
     "Object",
     "Time",
-    "azel",
+    "api",
     "compute",
     "consts",
     "get_location",
@@ -17,7 +17,7 @@ __version__ = "0.7.0"
 
 
 # submodules
-from . import azel
+from . import api
 from . import consts
 from . import location
 from . import object
@@ -26,7 +26,7 @@ from . import utils
 
 
 # aliases
-from .azel import compute
+from .api import compute
 from .location import Location, get_location
 from .object import Object, get_object
 from .time import Time, get_time
